plug-ins/sinus/sinus.c: applied Thomas' 1 liner to fix sinus for tilable.

+++ plug-ins/sinus/sinus.c
@@ -55,7 +55,7 @@
 #define RAND_MAX 2147483647
 #endif /* RAND_MAX */
 
-#define ROUND_TO_INT(val) ((val) + 0.5)
+#define ROUND_TO_INT(val) ((int) ((val) + 0.5))
 
 typedef gdouble colRGBA[4];
